risk_management/requirements.txt
python-binance
python-dotenv
pandas
numpy
python-telegram-bot
pytest
pytest-mock
yfinance
matplotlib
optuna